moved authorization middleware from routes/api/venues to utils/auth
